// src/types/zoho.ts
export type ZohoEventType =
  | 'Ticket_Add'
  | 'Ticket_Thread_Add'
  | 'Ticket_Update'
  | 'Ticket_Delete';
// ...
export interface ZohoTicketPayload {
  id?: string;
  ticketId?: string;
  ticketNumber: string;
  subject: string;
  description?: string;
  status: ZohoTicketStatus;
  priority: ZohoPriority;
  category?: string;
  subCategory?: string;
  channel: ZohoChannel;
  departmentId: string;
  productId?: string;
  contactId: string;
  contact: ZohoContact;
  email: string;
  assigneeId?: string;
  teamId?: string;
  dueDate?: string;
  closedTime?: string;
  createdTime: string;
  modifiedTime: string;
  webUrl: string;

  // First thread (for new tickets)
  firstThread?: ZohoThread;

  // Thread add specific
  content?: string;
  author?: ZohoAuthor;
  isForward?: boolean;
  to?: string;
}

// Webhook Body
export interface ZohoWebhookBody {
  eventType: ZohoEventType;
  orgId: string;
  payload: ZohoTicketPayload;
}

// Webhook Request
export interface ZohoWebhookRequest {
  headers: {
    host: string;
    'x-forwarded-for'?: string;
    'user-agent'?: string;
    'content-type': string;
  };
  body: ZohoWebhookBody[];
}